refactor(payment-service): use ConfigurationProperties for PayOS config

+++ backend/payment-service/src/main/java/com/CNTTK18/paymentservice/service/PaymentService.java
@@ -8,7 +8,7 @@ public interface PaymentService {
     /**
      * Luồng 1: Nhận yêu cầu tạo thanh toán từ Frontend -> Lưu Db -> Call PayOS -> Trả về Link
      */
-    String createPaymentLink(UUID userId, Integer amount);
+    String createPaymentLink(UUID userId, Long amount);
 
     /**
      * Luồng 2: Xử lý Webhook (thông báo) được bắn về từ PayOS sau khi KH quét mã thành công
